requirements.txt
Babel==2.12.1
backcall==0.2.0
Django==4.2.4
django-environ==0.11.0
django-modeltranslation==0.18.11
django-phonenumber-field==7.1.0
django-user-agents==0.4.0
pandas==2.1.0
phonenumbers==8.13.19
Pillow==10.0.0
user-agents==2.2.0

